♻️  research_bootstrap.py 추출 — collector + intake formatters (P0-L step 7)
변경 이유
- `_maybe_run_auto_collect` 호출부 가드 (P0-K command-only),
  `_format_intake_with_collection` / `_format_coding_bootstrap_body` /
  `_format_collection_*` 등 intake/collection 본문 포맷, 그리고
  ResearchCandidate 후보 분류 (`classify_attachment`/`classify_url`/
  `collect_research_candidates_from_message`/
  `suggest_role_research_assignments`/`format_insufficient_research_prompt`)
  까지 — 모두 "자료 수집 + intake 본문" 책임이라 한 모듈로 분리.

주요 변경 사항
- 새 파일 `engineering_conversation/research_bootstrap.py` (약 900 줄):
  - 11 SOURCE_TYPE 상수 + ALL_SOURCE_TYPES / IMAGE_EXTENSIONS /
    ROLE_RESEARCH_PROFILES / _REQUIRED_SOURCE_TYPES_BY_TASK_TYPE
  - host suffix 목록 3 종 (_DESIGN_REFERENCE_HOSTS /
    _OFFICIAL_DOCS_HOST_SUFFIXES / _COMMUNITY_SIGNAL_HOST_SUFFIXES)
  - `ResearchCandidate` / `ResearchCollectionResult` dataclass
  - 6 attachment helper (_attachment_field/_filename/_url/
    _content_type/_id/_size)
  - public 분류: `classify_attachment`, `classify_url`,
    `collect_research_candidates_from_message`,
    `suggest_role_research_assignments`,
    `format_insufficient_research_prompt`,
    `build_research_pack_from_candidates`
  - private 메타: `_why_relevant_for`, `_risk_or_limit_for`,
    `_confidence_for`, `_evaluate_research_sufficiency`,
    `_truncate`, `_url_title`
  - collector wiring: `_maybe_run_auto_collect`
  - intake 본문 4 종: `_format_intake_with_collection`,
    `_format_coding_bootstrap_body`,
    `_format_collection_announcement`, `_format_collection_meta_block`
- `_legacy.py` 의 해당 2 블록 (254 + 757 = 1,011 줄) 제거 후 re-export.
- `__init__.py` facade 가 research surface 도 `.research_bootstrap`
  직접 import 으로 끌어옴 — `_legacy` 의존이 거의 사라졌다.

비고
- 4657 PASS + 1 skipped 회귀 유지 (동작 변경 0).
- `_summarize_topic` / `_pretty_task_type` / `_pretty_provider` 은
  response_formatters 책임이라 research_bootstrap 안에서 lazy
  `from ._legacy import` 로 사용 — step 8 에서 정리.
- 다음 step (8): response_formatters.py 추출 후 `_legacy.py` 제거.
